Downloaded Subjects are now cached in SqlDelight Database.
Currently, the logic to download our subjects is still in the GetSubjects.kt use case.  Next step is to extract that out to a "startup" screen where we cache subjects and data from all of the other endpoints the first time the user opens the app. (Once we have our subjects, we will rarely ever have to update them, so the SubjectListScreen should always use the cache and we can validate the cache on startup/in the background with a WorkManager at regular intervals.)